requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
Django==3.2.4
djangorestframework==3.12.4
et-xmlfile==1.1.0
ezodf==0.3.2
lxml==4.6.3
numpy==1.21.0
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.0
pandas-ods-reader==0.0.7
postgres==3.0.0
psycopg2-binary==2.9.1
psycopg2-pool==1.1
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
